requirements.txt
flask
pillow
tweepy
python-dotenv
gunicorn
